Fixed some features
